refactor to facilitate implementation of main menu
changed how levels are loaded, there are two stages now. First levels
are loaded from the file and put in the asset list. Then, when a game
starts, selected level is copied to a game play state and then the
loading is finalized. This way, the levels can be more easily reused
without need for complex cleanup operation.